refactor: moves PanelHandler
